requirements.txt
fastapi
uvicorn 
groq 
python-dotenv
sqlmodel